Make repository handling of validations reusable
