+++ scripts/runValidation.sh
@@ -1,2 +1,2 @@
-java -cp "target/diffdetective-1.0.0-jar-with-dependencies.jar" org.variantsync.diffdetective.validation.Validation
+java -cp "target/diffdetective-1.0.0-jar-with-dependencies.jar" org.variantsync.diffdetective.validation.EditClassValidation
 echo "runValidation.sh DONE"
