+++ src/test/java/MarlinDebug.java
@@ -22,7 +22,7 @@ import org.variantsync.diffdetective.feature.CPPAnnotationParser;
 import org.variantsync.diffdetective.mining.DiffTreeMiner;
 import org.variantsync.diffdetective.editclass.proposed.ProposedEditClasses;
 import org.variantsync.diffdetective.util.Clock;
-import org.variantsync.diffdetective.validation.Validation;
+import org.variantsync.diffdetective.validation.EditClassValidation;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -148,7 +148,7 @@ public class MarlinDebug {
         final RevWalk revWalk = new RevWalk(git.getRepository());
         final RevCommit childCommit = revWalk.parseCommit(ObjectId.fromString(commitHash));
 
-        Validation.AnalysisFactory.apply(
+        EditClassValidation.AnalysisFactory.apply(
             repoInspection.repo,
             repoInspection.outputPath
         ).processCommits(List.of(childCommit));
